refactor: split and cleanup
